pyproject.toml
[project]
name = "meshmapper-webhook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.5",
    "argon2-cffi>=25.1.0",
    "beanie>=2.1.0",
    "fastapi[standard]>=0.135.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "ulid-py>=1.1.0",
]

[tool.fastapi]
entrypoint = "meshmapper_adapter:app"